# DatabaseService/bin/Release/net6.0/publish/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DatabaseService, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/List": {
      "get": {
        "tags": [
          "List"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToDoItems"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToDoItems"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToDoItems"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/List/List/Get/{id}": {
      "get": {
        "tags": [
          "List"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoItems"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoItems"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToDoItems"
                }
              }
            }
          }
        }
      }
    },
    "/List/List/AddItem": {
      "post": {
        "tags": [
          "List"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToDoItems"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ToDoItems"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ToDoItems"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/List/List/Delete/{idi}": {
      "delete": {
        "tags": [
          "List"
        ],
        "parameters": [
          {
            "name": "idi",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ToDoItems": {
        "type": "object",
        "properties": {
          "toDoItemsId": {
            "type": "string",
            "nullable": true
          },
          "details": {
            "type": "string",
            "nullable": true
          },
          "deadline": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}